Extract ManagedDependency into separate file
